+++ changelog.md
@@ -1,7 +1,7 @@
 ### v2.0.0 - 2020-08-??
  - foo_jscript_panel replaced with foo_spider_monkey_panel
  - Simplified script initialization (no more pasting contents of Georgia.txt into Configuration panel after initial setup time)
- - Automatically generating and reading preferences from georgia-preferences.json
+ - Automatically generating and reading preferences from georgia-config.jsonc
  - When using hyperlinks to search, if current playing song is in results, it will show as playing
  - Added georgia-config.jsonc file to store preferences outside foobar